agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qwickguard-agent"
version = "0.1.0"
description = "QwickGuard AI-powered self-healing infrastructure agent"
requires-python = ">=3.11"
dependencies = [
    "psutil>=5.9",
    "docker>=7.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
]

[project.scripts]
qwickguard-agent = "qwickguard_agent.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/qwickguard_agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
